fix/156/ move transactions tests into Streetcode directory and adjust namespaces
